+++ frontend/src/modules/purchasing/views/InventoryPriorityView.tsx
@@ -1,8 +1,0 @@
-import { memo } from "react";
-import PurchasingSegmentsPage from "../../../pages/purchasing/PurchasingSegmentsPage";
-
-function InventoryPriorityViewInner() {
-  return <PurchasingSegmentsPage />;
-}
-
-export default memo(InventoryPriorityViewInner);